Recognition/requirements.txt
Flask==2.2.2
Flask-API==3.0.post1
Flask-RESTful==0.3.9
opencv-python==4.6.0.66
Werkzeug==2.2.2
PyYAML==6.0
pytesseract==0.3.10
requests==2.31.0